fetch data added to app js
